shadow/openssh: use pam_selinux only if target_selinux enabled
We add pam conf files for login/sshd to use pam_selinux module. When
selinux is not in DISTRO_FEATURES, pam-plugin-selinux would not be
built, this will cause runtime errors to not allow users to login in
on the console or ssh.
Use @target_selinux() to enable these pam conf files conditionally.

Signed-off-by: Xin Ouyang <Xin.Ouyang@windriver.com>